Reimplement lexer and parser in Parslet
There was a lot of nasty ad-hoc state-machine malarky in the lexer and
parser. This replaces it with a parser written in Parslet, which is a
fairly nice PEG parser generator.

See: https://kschiess.github.io/parslet/

This should make it easier to implement floats, strings, and other
stuff in the future.

Note that we're using a Parslet transformer to strip out the metadata
tags that Parslet adds to the AST for us (stuff like :sexpr, :integer,
etc). We'll probably want to add this in again once we start thinking
a little harder about typing, but right now we just want it to work
with the evaluator we've already got.