Refactor environment frames, fix closures
It turns out that the way we were managing closures was totally
busted, since the stack was essentially global to the evaluator. That
definitely doesn't work; closures don't work that way. This new system
(centered around frames, not "environments") should address that
problem. It also simplifies the whole notion of scoping considerably.

A downside is that it (currently) busts the way the load function
works, so I've made that a special form again for now.