Rename "macros" to "fexprs"
The terms that Blueprint calls macros are expanded at runtime (rather
than in a separate macro-expansion phase). It's not exactly wrong to
call them macros, but the more precise term would be "fexprs" and we
should refer to them as such in the code and documentation.

On the topic of fexprs, Kent Pitman wrote "Special Forms in Lisp" in
1980 recommending against their use because they aren't amenable to
static analysis. Blueprint performs exactly no static analysis, and
doesn't aspire to, so this isn't a big problem for us.

See also:
- https://en.wikipedia.org/wiki/Fexpr
- http://www.nhplace.com/kent/Papers/Special-Forms.html